Refactoring. Ability to choose the age and city of the interlocutor. Ability to choose your own age and city. Ability to configure Polling. If webhook is invalid, the bot starts using "Polling".

+++ db/db.go
@@ -9,7 +9,7 @@ import (
 
 type database interface {
 	SaveUser(user betypes.User) error
-	GetUser(userID int64) (*betypes.User, error)
+	GetUser(userID int) (*betypes.User, error)
 }
 
 // RedisDB saves redis database settings.
